requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib
jinja2
python-multipart